initial commit
development version cloning to local only first.

using git config and resource file to store last project.

.wp-sandbox/index.php file for quick info

using the wordpress docker container package.

and having  https://github.com/docker-library/wordpress/issues/333
which is a pity, but the extra-config.php script for the rescue.

and possbile to inject xdebug on the fly. this costs a bit, should
maybe keep the .so files to spare the compile time each time, WIP.

add some docs, yet missing the manpage.